+++ src/components/search/sorting-selector/src/SortingSelector.tsx
@@ -1,4 +1,5 @@
 import * as React from "react";
+import * as PropTypes from "prop-types";
 
 import {
 	SearchkitComponent,
@@ -29,12 +30,12 @@ export class SortingSelector extends SearchkitComponent<SortingProps, any> {
 
   static propTypes = defaults({
     listComponent: RenderComponentPropType,
-    options:React.PropTypes.arrayOf(
-      React.PropTypes.shape({
-        label:React.PropTypes.string.isRequired,
-        field:React.PropTypes.string,
-        order:React.PropTypes.string,
-        defaultOption:React.PropTypes.bool
+    options:PropTypes.arrayOf(
+      PropTypes.shape({
+        label:PropTypes.string.isRequired,
+        field:PropTypes.string,
+        order:PropTypes.string,
+        defaultOption:PropTypes.bool
       })
     )
   }, SearchkitComponent.propTypes)
@@ -58,7 +59,7 @@ export class SortingSelector extends SearchkitComponent<SortingProps, any> {
 
   render() {
     const { listComponent } = this.props
-		const options = this.accessor.options.options		
+		const options = this.accessor.options.options
     const selected = [this.accessor.getSelectedOption().key]
     const disabled = !this.hasHits()